Fix invalid memory access with invalid powers
The ParseNode's value must be valid to pass to `free()`, as it's done
unconditionally if the expression is invalid.  However, for some
functions a shortuct was taken avoiding duplicating memory.  This was
OK in case the expression is valid because the evaluation function
would convert it and set the pointer back to NULL; but it's not OK if
the expression is invalid as the evaluation doesn't happen and the
pointer is not reset before being freed.

Fixes #226.